Refactor for single cron, multiple function architecture to comply with hobby vercel tier limits
