+++ src/lib/email/lifecycle/index.ts
@@ -1,0 +1,9 @@
+export type LifecycleResult = {
+  name: string;
+  candidates: number;
+  sent: number;
+  skipped: number;
+  failed: number;
+};
+
+export { runProfileCompletionReminders } from "./profileCompletionReminder";
