+++ vercel.json
@@ -1,7 +1,7 @@
 {
   "crons": [
     {
-      "path": "/api/cron/send-profile-completion-reminders",
+      "path": "/api/cron/lifecycle-emails",
       "schedule": "0 3 * * *"
     }
   ]
